Bloc-6/Api-Pneumonia/requirements.txt
flask==2.0.2
keras==2.8.0
numpy==1.22.2
pandas==1.4.0
pillow==9.0.1